Add clean up function for public dir

+++ service/AppService.js
@@ -1,28 +1,18 @@
 const fs = require("fs");
-const { errorResponse, successResponse } = require("../domain/Response");
-const SocketMessages = require("../socket-messages");
-const { decodeJWT } = require("../utils");
+const { Time } = require("../utils/constants");
 
 const AppService = {
-  sendMessages: async (token) => {
-    try {
-      if (decodeJWT(token)) {
-        return successResponse(SocketMessages);
-      }
-    } catch (error) {
-      return errorResponse("Something went wrong");
-    }
-  },
-  cleanupPublicDir: () => {
-    setInterval(async () => {
-      const files = (await fs.promises.readdir("./public")).filter(
-        (file) => file !== ".keep"
-      );
-      for (const file of files) {
+  cleanupPublicDir: async () => {
+    const files = (await fs.promises.readdir("./public")).filter(
+      (file) => file !== ".keep"
+    );
+    for (const file of files) {
+      const { birthtime } = fs.statSync(`./public/${file}`);
+      if (new Date().getTime() - birthtime.getTime() > Time.ONE_DAY) {
         await fs.promises.rm(`./public/${file}`);
         console.log(`Cleaned ./public/${file}`);
       }
-    }, 1000 * 3600 * 24);
+    }
     return true;
   },
 };
